requirements.txt
Flask==3.1.3
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
Werkzeug==3.1.5
python-dotenv==1.2.1
pdf2docx==0.5.9
pdfplumber==0.11.9
openpyxl==3.1.5
docx2pdf==0.1.8
mammoth==1.11.0
reportlab==4.4.10
gunicorn==25.1.0
